requirement.txt
Flask==2.3.3
numpy==1.24.4
opencv-python==4.8.0.76
pillow==9.4.0
gunicorn